proper hashmap internal vector behavior on replace
json_hmap_t stores a json_vec_t internally in order to keep pairs
ordered. when calling put with a duplicate key, hmap would replace the
old hashmap node but still push the key on the vec, resulting in
duplicate output when serializing. now hmap_put checks for this
condition